Load hack from sever.
Hacks can now be loaded from the server using the query string param "hackid".
The response from the server when saving a hack is now used to refresh the state of the editor.
Ignored the jshack-common-web application root directory so that hacks created during testing don't get added to git. Consider moving this directory to the build directory.