Begining of implementation of file upload.
Hit an issue with Jersey 2.4.1 which is going to be resolved in v2.5: https://java.net/jira/browse/JERSEY-2197
This is going to be a bit of a problem as it pretty much breaks Jersey completely when the resource classes are inside a jar within a war, which is the case in JS Hack. It will not read past the first resource.